+++ idemio/benches/route_benchmarks.rs
@@ -6,7 +6,8 @@ use idemio::handler::{Handler, HandlerId};
 use idemio::router::config::builder::{
     MethodBuilder, RouteBuilder, ServiceBuilder, SingleServiceConfigBuilder,
 };
-use idemio::router::path::{PathMatcher, PathPrefixMethodKey, PathPrefixMethodPathMatcher};
+use idemio::router::path::PathMatcher;
+use idemio::router::path::http::PathPrefixMethodPathMatcher;
 use idemio::status::{ExchangeState, HandlerStatus};
 use std::convert::Infallible;
 use std::hint::black_box;
@@ -29,7 +30,9 @@ impl Handler<(), (), ()> for DummyHandler {
     }
 }
 
-fn create_populated_dynamic_route_table_v2(num_routes: usize) -> PathPrefixMethodPathMatcher<(), (), ()> {
+fn create_populated_dynamic_route_table_v2(
+    num_routes: usize,
+) -> PathPrefixMethodPathMatcher<(), (), ()> {
     let mut registry = HandlerRegistry::new();
     registry
         .register_handler(HandlerId::new("test1"), DummyHandler)
