+++ idemio/src/router/factory.rs
@@ -134,7 +134,7 @@ pub mod hyper {
     use hyper::http::request::Parts;
     use std::error::Error;
     use std::marker::PhantomData;
-    use crate::router::path::PathPrefixMethodKey;
+    use crate::router::path::http::PathPrefixMethodKey;
 
     /// Factory implementation for creating exchanges from Hyper HTTP requests.
     ///
